requirements.txt
accelerate==1.6.0
amqp==5.3.1
appdirs==1.4.4
argostranslate==1.9.6
asgiref==3.8.1
attrs==25.3.0
backoff==2.2.1
billiard==4.2.1
bitsandbytes==0.45.5
blobfile==3.0.0
boilerpy3==1.0.7
cattrs==24.1.3
celery==5.5.2
certifi==2025.4.26
chardet==3.0.4
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
ctranslate2==4.6.0
DAWG-Python==0.7.2
distro==1.9.0
Django==5.2.1
django-crispy-forms==2.4
django-extensions==4.1
docopt==0.6.2
Events==0.5
exceptiongroup==1.2.2
farm-haystack==1.26.4.post0
filelock==3.18.0
fsspec==2025.3.2
ftfy==6.3.1
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hf-xet==1.1.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.31.1
hyperframe==5.2.0
idna==2.10
inflect==7.5.0
Jinja2==3.1.6
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.5.3
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
lazy-imports==0.3.1
llvmlite==0.44.0
locate==1.1.1
lxml==5.4.0
marisa-trie==1.2.1
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
nltk==3.9.1
num2words==0.5.14
numba==0.61.2
numpy==2.2.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
openai-whisper==20240930
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
posthog==4.0.1
prompt_toolkit==3.0.51
prompthub-py==4.0.0
protobuf==6.30.2
psutil==7.0.0
pycryptodomex==3.22.0
pydantic==1.10.22
pydub==0.25.1
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
quantulum3==0.9.2
rank-bm25==0.2.2
razdel==0.5.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-cache==0.9.8
rfc3986==1.5.0
rpds-py==0.24.0
sacremoses==0.0.53
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sseclient-py==1.8.0
stanza==1.1.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
transformers==4.51.3
triton==3.0.0
typeguard==4.4.2
typing_extensions==4.13.2
tzdata==2025.2
url-normalize==1.4.3
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
wordfreq==3.1.1
